+++ lib/searchkick/reindex_v2_job.rb
@@ -19,6 +19,5 @@ module Searchkick
         index.store record
       end
     end
-
   end
 end