+++ lib/searchkick/results.rb
@@ -15,27 +15,24 @@ module Searchkick
       @options = options
     end
 
+    # experimental: may not make next release
+    def records
+      @records ||= results_query(klass, hits)
+    end
+
     def results
       @results ||= begin
         if options[:load]
           # results can have different types
           results = {}
 
-          hits.group_by { |hit, i| hit["_type"] }.each do |type, grouped_hits|
-            records = type.camelize.constantize
-            if options[:includes]
-              if defined?(NoBrainer::Document) && records < NoBrainer::Document
-                records = records.preload(options[:includes])
-              else
-                records = records.includes(options[:includes])
-              end
-            end
-            results[type] = results_query(records, grouped_hits)
+          hits.group_by { |hit, _| hit["_type"] }.each do |type, grouped_hits|
+            results[type] = results_query(type.camelize.constantize, grouped_hits).to_a.index_by { |r| r.id.to_s }
           end
 
           # sort
           hits.map do |hit|
-            results[hit["_type"]].find { |r| r.id.to_s == hit["_id"].to_s }
+            results[hit["_type"]][hit["_id"].to_s]
           end.compact
         else
           hits.map do |hit|
@@ -76,6 +73,23 @@ module Searchkick
 
     def facets
       response["facets"]
+    end
+
+    def aggs
+      @aggs ||= begin
+        response["aggregations"].dup.each do |field, filtered_agg|
+          buckets = filtered_agg[field]
+          # move the buckets one level above into the field hash
+          if buckets
+            filtered_agg.delete(field)
+            filtered_agg.merge!(buckets)
+          end
+        end
+      end
+    end
+
+    def took
+      response["took"]
     end
 
     def model_name
@@ -141,19 +155,30 @@ module Searchkick
 
     private
 
-    def results_query(records, grouped_hits)
+    def results_query(records, hits)
+      ids = hits.map { |hit| hit["_id"] }
+
+      if options[:includes]
+        records =
+          if defined?(NoBrainer::Document) && records < NoBrainer::Document
+            records.preload(options[:includes])
+          else
+            records.includes(options[:includes])
+          end
+      end
+
       if records.respond_to?(:primary_key) && records.primary_key
         # ActiveRecord
-        records.where(records.primary_key => grouped_hits.map { |hit| hit["_id"] }).to_a
+        records.where(records.primary_key => ids)
       elsif records.respond_to?(:all) && records.all.respond_to?(:for_ids)
         # Mongoid 2
-        records.all.for_ids(grouped_hits.map { |hit| hit["_id"] }).to_a
+        records.all.for_ids(ids)
       elsif records.respond_to?(:queryable)
         # Mongoid 3+
-        records.queryable.for_ids(grouped_hits.map { |hit| hit["_id"] }).to_a
+        records.queryable.for_ids(ids)
       elsif records.respond_to?(:unscoped) && records.all.respond_to?(:preload)
         # Nobrainer
-        records.unscoped.where(:id.in => grouped_hits.map { |hit| hit["_id"] }).to_a
+        records.unscoped.where(:id.in => ids)
       else
         raise "Not sure how to load records"
       end