+++ test/ci/before_install.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 
-wget https://download.elasticsearch.org/elasticsearch/elasticsearch/elasticsearch-1.3.2.deb
-sudo dpkg -i elasticsearch-1.3.2.deb
-sudo service elasticsearch restart
+sudo apt-get purge elasticsearch
+wget https://download.elasticsearch.org/elasticsearch/elasticsearch/elasticsearch-1.7.3.deb
+sudo dpkg -i elasticsearch-1.7.3.deb
+sudo service elasticsearch start
 
 if [ -n "$NOBRAINER" ]; then
   source /etc/lsb-release && echo "deb http://download.rethinkdb.com/apt $DISTRIB_CODENAME main" | sudo tee /etc/apt/sources.list.d/rethinkdb.list