+++ test/model_test.rb
@@ -1,7 +1,6 @@
 require_relative "test_helper"
 
-class TestModel < Minitest::Test
-
+class ModelTest < Minitest::Test
   def test_disable_callbacks_model
     store_names ["product a"]
 
@@ -20,7 +19,7 @@ class TestModel < Minitest::Test
   def test_disable_callbacks_global
     # make sure callbacks default to on
     assert Searchkick.callbacks?
-    
+
     store_names ["product a"]
 
     Searchkick.disable_callbacks
@@ -34,5 +33,4 @@ class TestModel < Minitest::Test
 
     assert_search "product", ["product a", "product b"]
   end
-
 end